Add documentation lookup (⌃H)
This command uses `docsetutil` to lookup the current word/selection.

It does not filter out C or Objective-C results since some things are only available in these docs. It also does not filter on iOS/OS X since we don’t know which of the SDKs the user prefer.

The command borrows heavily from the documentation lookup command in the Objective-C bundle, it would make sense to merge them, but the version in the Objective-C bundle also looks for selectors and manual pages so it is a tad more complex.