Treat `xcrun --sdk macosx swift` as a single command
When running a script containing a correct hashbang, like

    #!/usr/bin/env xcrun --sdk macosx swift

the TextMate::Executor would replace the *first* element of `command`
with the parsed hashbang command.

This results in TextMate trying to execute commands like

    /usr/bin/env xcrun --sdk macosx swift --sdk macosx swift

which fails because swift(1) doesn’t accept an `--sdk` parameter.

This change specifies the whole `xcrun … swift` command as a single
element, which it is anyway conceptually, and which fixes this problem.