Match opaque result types per SE-0244 (#44)
Documentation:
- https://docs.swift.org/swift-book/ReferenceManual/Types.html#ID616
- https://github.com/apple/swift-evolution/blob/master/proposals/0244-opaque-result-types.md

Opaque result types are used heavily in SwiftUI code, for example:
- https://github.com/ivanvorobei/SwiftUI
- https://github.com/apple-sample-code/SwiftUI-Tutorials

Since `some` is a keyword that pertains only to types, it can still be used as a regular variable name. So this PR adds a pattern for `let`/`var` declarations with types, so that the `some` pattern can be enabled only where appropriate.